+++ doc/changelog.md
@@ -186,3 +186,4 @@ v0.7.0
 + (done, tested) Add configurations: subsite-configs & subsite-modules (JSON Format) for default installation on register
 + (done, tested) Add setting for third party authentication
 + (done, tested) Change registration, let registrant upload favicon, logo, set slogan & site name
++ (proposed) Use this `$_SERVER['HTTP_HOST']`